+++ geofileops/geoops.py
@@ -11,7 +11,7 @@ from typing import Any, Callable, List, Literal, Optional, Tuple, Union, TYPE_CH
 import warnings
 
 from pygeoops import GeometryType
-import shapely
+import pygeoops
 
 from geofileops import fileops
 from geofileops.util import _geoops_gpd
@@ -1158,7 +1158,7 @@ def makevalid(
     output_layer: Optional[str] = None,
     columns: Optional[List[str]] = None,
     explodecollections: bool = False,
-    force_output_geometrytype: Optional[GeometryType] = None,
+    force_output_geometrytype: Union[str, None, GeometryType] = None,
     gridsize: float = 0.0,
     keep_empty_geoms: Optional[bool] = None,
     where_post: Optional[str] = None,
@@ -1175,7 +1175,7 @@ def makevalid(
 
     Alternative names:
         - QGIS: fix geometries
-        - shapely: make_valid
+        - shapely, geopandas: make_valid
 
     If ``explodecollections`` is False and the output file is a GeoPackage, the fid
     will be preserved. In other cases this will typically not be the case.
@@ -1194,7 +1194,7 @@ def makevalid(
         explodecollections (bool, optional): True to output only simple geometries.
             Defaults to False.
         force_output_geometrytype (GeometryType, optional): The output geometry type to
-            force the output to. If None, the geometry type of the input is used.
+            force the output to. If None, the geometry type of the input is retained.
             Defaults to None.
         gridsize (float, optional): the size of the grid the coordinates of the ouput
             will be rounded to. Eg. 0.001 to keep 3 decimals. Value 0.0 doesn't change
@@ -1240,10 +1240,25 @@ def makevalid(
             stacklevel=2,
         )
 
+    # Determine if collapsed parts need to be kept after makevalid or not
+    keep_collapsed = True
+    if force_output_geometrytype is None:
+        keep_collapsed = False
+    else:
+        if isinstance(force_output_geometrytype, GeometryType):
+            force_output_geometrytype = force_output_geometrytype.name
+        info = fileops.get_layerinfo(input_path)
+        if force_output_geometrytype.startswith(
+            info.geometrytypename
+        ) or info.geometrytypename.startswith(force_output_geometrytype):
+            keep_collapsed = False
+
     _geoops_gpd.apply(
         input_path=Path(input_path),
         output_path=Path(output_path),
-        func=lambda geom: shapely.make_valid(geom),
+        func=lambda geom: pygeoops.make_valid(
+            geom, keep_collapsed=keep_collapsed, only_if_invalid=True
+        ),
         operation_name="makevalid",
         input_layer=input_layer,
         output_layer=output_layer,
@@ -2347,6 +2362,8 @@ def join_nearest(
     input2_path: Union[str, "os.PathLike[Any]"],
     output_path: Union[str, "os.PathLike[Any]"],
     nb_nearest: int,
+    distance: Optional[float] = None,
+    expand: Optional[bool] = None,
     input1_layer: Optional[str] = None,
     input1_columns: Optional[List[str]] = None,
     input1_columns_prefix: str = "l1_",
@@ -2359,7 +2376,23 @@ def join_nearest(
     force: bool = False,
 ):
     """
-    Joins features with the nb_nearest features that are closest.
+    Joins features of `input1` with the `nb_nearest` closest features of `input2`.
+
+    In addition to the columns requested via the `input*_columns` parameters, the
+    following columns will be in the output file as well:
+        - pos (int): relative rank (sorted by distance): the closest item will be #1,
+          the second closest item will be #2 and so on.
+        - distance (float): if the dataset is in a planar (= projected) crs, `distance`
+          will be in the unit defined by the projection (meters, feet, chains etc.).
+          For a geographic dataset (longitude and latitude degrees), `distance` will be
+          in meters, with the most precise geodetic formulas being applied.
+        - distance_crs (float): if the dataset is in a planar (= projected) crs,
+          `distance_crs` will be in the unit defined by the projection (meters, feet,
+          chains etc.). For a geographic dataset (longitude and latitude degrees),
+          `distance_crs` will be in angles. Only available with spatialite >= 5.1.
+
+    Note: if spatialite version >= 5.1 is used, parameters `distance` and `expand` are
+    mandatory.
 
     Args:
         input1_path (PathLike): the input file to join to nb_nearest features.
@@ -2367,8 +2400,17 @@ def join_nearest(
         output_path (PathLike): the file to write the result to
         nb_nearest (int): the number of nearest features from input 2 to join
             to input1.
+        distance (float): maximum distance to search for the nearest items. If `expand`
+            is True, this is the initial search distance, which will be gradually
+            expanded (doubled) till `nb_nearest` are found. For optimal performance,
+            it is important to choose the typical value that will be needed to find
+            `nb_nearest` items. If `distance` is too large, performance can be bad.
+            Parameter is only relevant if spatialite version >= 5.1 is used.
+        expand (bool): `True` to keep searching till nb_nearest items are found. If
+            `False`, only items found within `distance` are returned (`False` is only
+            supported if spatialite version >= 5.1 is used).
         input1_layer (str, optional): input layer name. Optional if the
-            file only contains one layer. Defaults to None.
+            file only contains one layer. Defaults to `None`.
         input1_columns (List[str], optional): list of columns to retain. If None, all
             standard columns are retained. In addition to standard columns, it is also
             possible to specify "fid", a unique index available in all input files. Note
@@ -2401,6 +2443,8 @@ def join_nearest(
         input2_path=Path(input2_path),
         output_path=Path(output_path),
         nb_nearest=nb_nearest,
+        distance=distance,
+        expand=expand,
         input1_layer=input1_layer,
         input1_columns=input1_columns,
         input1_columns_prefix=input1_columns_prefix,